+++ src/OneVarConstraint.cpp
@@ -48,10 +48,10 @@ void OneVarConstraint::set_variable( ColVariable * const variable,
  if( f_variable )   // add this Constraint to its only active Variable
   f_variable->add_active( this );
 
- if( ( ! f_Block ) || ( ! f_Block->issue_mod( issueMod ) ) )
+ if( ( ! get_Block() ) || ( ! get_Block()->issue_mod( issueMod ) ) )
   return;
 
- f_Block->add_Modification( std::make_shared< OneVarConstraintMod >(
+ get_Block()->add_Modification( std::make_shared< OneVarConstraintMod >(
                              this , OneVarConstraintMod::eVariableChanged ,
                              Observer::par2concern( issueMod ) ) ,
                             Observer::par2chnl( issueMod ) );
@@ -67,10 +67,10 @@ void BoxConstraint::set_rhs( c_RHSValue rhs_value , ModParam issueMod )
 
  f_rhs = rhs_value;        // change the value
 
- if( ( ! f_Block ) || ( ! f_Block->issue_mod( issueMod ) ) )
+ if( ( ! get_Block() ) || ( ! get_Block()->issue_mod( issueMod ) ) )
   return;
 
- f_Block->add_Modification( std::make_shared< OneVarConstraintMod >(
+ get_Block()->add_Modification( std::make_shared< OneVarConstraintMod >(
                              this , RowConstraintMod::eChgRHS ,
                              Observer::par2concern( issueMod ) ) ,
                             Observer::par2chnl( issueMod ) );
@@ -85,10 +85,10 @@ void BoxConstraint::set_lhs( c_RHSValue lhs_value , ModParam issueMod )
 
  f_lhs = lhs_value;        // change the value
 
- if( ( ! f_Block ) || ( ! f_Block->issue_mod( issueMod ) ) )
+ if( ( ! get_Block() ) || ( ! get_Block()->issue_mod( issueMod ) ) )
   return;
 
- f_Block->add_Modification( std::make_shared< OneVarConstraintMod >(
+ get_Block()->add_Modification( std::make_shared< OneVarConstraintMod >(
                              this , RowConstraintMod::eChgLHS ,
                              Observer::par2concern( issueMod ) ) ,
                             Observer::par2chnl( issueMod ) );
@@ -104,10 +104,10 @@ void BoxConstraint::set_both( c_RHSValue both_value , ModParam issueMod )
  f_lhs = both_value;
  f_rhs = both_value;
 
- if( ( ! f_Block ) || ( ! f_Block->issue_mod( issueMod ) ) )
+ if( ( ! get_Block() ) || ( ! get_Block()->issue_mod( issueMod ) ) )
   return;
 
- f_Block->add_Modification( std::make_shared< OneVarConstraintMod >(
+ get_Block()->add_Modification( std::make_shared< OneVarConstraintMod >(
                              this , RowConstraintMod::eChgBTS ,
                              Observer::par2concern( issueMod ) ) ,
                             Observer::par2chnl( issueMod ) );
@@ -122,10 +122,10 @@ void LB0Constraint::set_rhs( c_RHSValue rhs_value, ModParam issueMod )
 
  f_rhs = rhs_value;        // change the value
 
- if( ( ! f_Block ) || ( ! f_Block->issue_mod( issueMod ) ) )
+ if( ( ! get_Block() ) || ( ! get_Block()->issue_mod( issueMod ) ) )
   return;
 
- f_Block->add_Modification( std::make_shared< OneVarConstraintMod >(
+ get_Block()->add_Modification( std::make_shared< OneVarConstraintMod >(
                              this , RowConstraintMod::eChgRHS ,
                              Observer::par2concern( issueMod ) ) ,
                             Observer::par2chnl( issueMod ) );
@@ -140,10 +140,10 @@ void UB0Constraint::set_lhs( c_RHSValue lhs_value , ModParam issueMod )
 
  f_lhs = lhs_value;        // change the value
 
- if( ( ! f_Block ) || ( ! f_Block->issue_mod( issueMod ) ) )
+ if( ( ! get_Block() ) || ( ! get_Block()->issue_mod( issueMod ) ) )
   return;
 
- f_Block->add_Modification( std::make_shared< OneVarConstraintMod >(
+ get_Block()->add_Modification( std::make_shared< OneVarConstraintMod >(
                              this , RowConstraintMod::eChgLHS ,
                              Observer::par2concern( issueMod ) ) ,
                             Observer::par2chnl( issueMod ) );
@@ -158,10 +158,10 @@ void LBConstraint::set_lhs( c_RHSValue lhs_value , ModParam issueMod )
 
  f_lhs = lhs_value;        // change the value
 
- if( ( ! f_Block ) || ( ! f_Block->issue_mod( issueMod ) ) )
+ if( ( ! get_Block() ) || ( ! get_Block()->issue_mod( issueMod ) ) )
   return;
 
- f_Block->add_Modification( std::make_shared< OneVarConstraintMod >(
+ get_Block()->add_Modification( std::make_shared< OneVarConstraintMod >(
                              this , RowConstraintMod::eChgLHS ,
                              Observer::par2concern( issueMod ) ) ,
                             Observer::par2chnl( issueMod ) );
@@ -176,10 +176,10 @@ void UBConstraint::set_rhs( c_RHSValue rhs_value, ModParam issueMod )
 
  f_rhs = rhs_value;        // change the value
 
- if( ( ! f_Block ) || ( ! f_Block->issue_mod( issueMod ) ) )
+ if( ( ! get_Block() ) || ( ! get_Block()->issue_mod( issueMod ) ) )
   return;
 
- f_Block->add_Modification( std::make_shared< OneVarConstraintMod >(
+ get_Block()->add_Modification( std::make_shared< OneVarConstraintMod >(
                              this , RowConstraintMod::eChgRHS ,
                              Observer::par2concern( issueMod ) ) ,
                             Observer::par2chnl( issueMod ) );